CNS-239: moved block time calc to different file
